refactor: decompose JointedModelImp (955→463 lines) into three managers
Extract joint hierarchy, resource binding, and visual management from
JointedModelImp into dedicated package-private classes:

- JointedModelResourceBinder: factory delegation, reflection-based joint
  ID discovery, isSims(), resource queries (134 lines)
- JointedModelVisualManager: VisualData lifecycle, scale properties,
  visual appearance management (135 lines)
- JointHierarchyManager: JointImpWrapper, joint maps, tree walk, IK
  chain computation, pose/straighten operations (652 lines)

JointedModelImp retains all public/protected API methods as thin
delegations, preserving full backward compatibility. All 3 inner
interfaces (VisualData, JointImplementationAndVisualDataFactory,
TreeWalkObserver) remain on JointedModelImp.

All 165 tests pass including 35 new decomposition contract tests.
