+++ core/story-api/src/main/java/org/lgna/story/implementation/JointedModelImp.java
@@ -46,31 +46,22 @@ package org.lgna.story.implementation;
 import edu.cmu.cs.dennisc.animation.Animated;
 import edu.cmu.cs.dennisc.animation.DurationBasedAnimation;
 import edu.cmu.cs.dennisc.animation.Style;
-import edu.cmu.cs.dennisc.java.lang.reflect.ReflectionUtilities;
-import edu.cmu.cs.dennisc.java.util.Lists;
-import edu.cmu.cs.dennisc.java.util.Maps;
-import edu.cmu.cs.dennisc.java.util.logging.Logger;
 import edu.cmu.cs.dennisc.math.EpsilonUtilities;
 import edu.cmu.cs.dennisc.property.InstanceProperty;
 import edu.cmu.cs.dennisc.property.event.PropertyListener;
-import edu.cmu.cs.dennisc.render.gl.imp.adapters.AdapterFactory;
 import edu.cmu.cs.dennisc.scenegraph.*;
 import edu.cmu.cs.dennisc.scenegraph.bound.CumulativeBound;
 import org.alice.math.immutable.*;
 import org.lgna.ik.core.solver.Bone;
-import org.lgna.ik.core.solver.Bone.Direction;
 import org.lgna.story.Paint;
 import org.lgna.story.Pose;
-import org.lgna.story.SJoint;
 import org.lgna.story.SJointedModel;
 import org.lgna.story.implementation.visualization.JointedModelVisualization;
-import org.lgna.story.resources.DynamicResource;
 import org.lgna.story.resources.JointArrayId;
 import org.lgna.story.resources.JointId;
 import org.lgna.story.resources.JointedModelResource;
 
-import java.lang.reflect.Field;
-import java.util.*;
+import java.util.List;
 
 /**
  * @author Dennis Cosgrove
@@ -106,307 +97,73 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     boolean isSims();
   }
 
-  private class JointImpWrapper extends JointImp {
-    public JointImpWrapper(JointedModelImp<?, ?> jointedModelImp, JointImp joint) {
-      super(jointedModelImp);
-      this.internalJointImp = joint;
+  public interface TreeWalkObserver {
+    void pushJoint(JointImp joint);
 
-    }
+    void handleBone(JointImp parent, JointImp child);
 
-    @Override
-    public final void setAbstraction(SJoint abstraction) {
-      super.setAbstraction(abstraction);
-      if (this.internalJointImp != null) {
-        this.internalJointImp.setAbstraction(abstraction);
-      }
-    }
-
-    @Override
-    public JointImp getJointParent() {
-      return jointParentWrapper;
-    }
-
-    @Override
-    public List<JointImp> getJointChildren() {
-      return jointChildrenWrapper;
-    }
-
-    @Override
-    void setJointParent(JointImp jointParent) {
-      if (this.jointParentWrapper != null) {
-        this.jointParentWrapper.getJointChildren().remove(this);
-      }
-      this.jointParentWrapper = jointParent;
-      if (this.jointParentWrapper != null) {
-        jointParent.getJointChildren().add(this);
-      }
-    }
-
-    @Override
-    public void setScale(Dimension3 scale) {
-      internalJointImp.setScale(scale);
-    }
-
-    @Override
-    public boolean isReoriented() {
-      return internalJointImp.isReoriented();
-    }
-
-    @Override
-    public boolean isRelocated() {
-      return internalJointImp.isRelocated();
-    }
-
-    @Override
-    protected void copyOnto(JointImp newJoint) {
-      internalJointImp.copyOnto(newJoint);
-      if (JointedModelImp.this.factory.isSims()) {
-        // Alice models reuse joints, but Sims regenerate them.
-        // Creating the adapter is the public mechanism to rebuild internal listeners and show joint changes
-        AdapterFactory.getAdapterFor(newJoint.getSgComposite());
-      }
-      if (getAbstraction() != null) {
-        newJoint.setAbstraction(getAbstraction());
-      }
-    }
-
-    @Override
-    public String getName() {
-      return internalJointImp.getJointId().toString();
-    }
-
-    @Override
-    public SceneImp getScene() {
-      return this.internalJointImp.getScene();
-    }
-
-    @Override
-    public JointId getJointId() {
-      return internalJointImp.getJointId();
-    }
-
-    @Override
-    public boolean isFreeInX() {
-      return internalJointImp.isFreeInX();
-    }
-
-    @Override
-    public boolean isFreeInY() {
-      return internalJointImp.isFreeInY();
-    }
-
-    @Override
-    public boolean isFreeInZ() {
-      return internalJointImp.isFreeInZ();
-    }
-
-    void replaceWithJoint(JointImp newJoint) {
-      copyOnto(newJoint);
-      AbstractTransformable oldSgComposite = internalJointImp.getSgComposite();
-      AbstractTransformable newSgComposite = newJoint.getSgComposite();
-      for (Component child : oldSgComposite.getComponents()) {
-        if (!(child instanceof ModelJoint)) {
-          child.setParent(newJoint.getSgComposite());
-        }
-      }
-      // Sims models, with regenerated joints, need to be reconnected at their root, indicated by a null parent.
-      if (newSgComposite.getParent() == null) {
-        // Without this, things riding on the Sim or its joints when the resource changes will go out of the scene graph and disappear.
-        newSgComposite.setParent(oldSgComposite.getParent());
-      }
-      internalJointImp = newJoint;
-    }
-
-    @Override
-    public AxisAlignedBox getAxisAlignedMinimumBoundingBox(ReferenceFrame asSeenBy) {
-      return internalJointImp.getAxisAlignedMinimumBoundingBox(asSeenBy);
-    }
-
-    @Override
-    public AbstractTransformable getSgComposite() {
-      return internalJointImp.getSgComposite();
-    }
-
-    @Override
-    protected void updateCumulativeBound(CumulativeBound rv, AffineMatrix4x4 trans) {
-      internalJointImp.updateCumulativeBound(rv, trans);
-    }
-
-    @Override
-    public UnitQuaternion getOriginalOrientation() {
-      return internalJointImp.getOriginalOrientation();
-    }
-
-    @Override
-    public AffineMatrix4x4 getScaledOriginalTransformation() {
-      return internalJointImp.getScaledOriginalTransformation();
-    }
-
-    @Override
-    public AffineMatrix4x4 getLocalTransformation() {
-      return internalJointImp.getLocalTransformation();
-    }
-
-    @Override
-    public void setLocalTransformation(AffineMatrix4x4 transformation) {
-      internalJointImp.setLocalTransformation(transformation);
-    }
-
-    @Override
-    protected void postCheckSetVehicle(EntityImp vehicle) {
-      //todo?
-      internalJointImp.postCheckSetVehicle(vehicle);
-    }
-
-    @Override
-    public boolean isFacing(EntityImp other) {
-      return this.internalJointImp.isFacing(other);
-    }
-
-    @Override
-    public void applyTranslation(double x, double y, double z, ReferenceFrame asSeenBy) {
-      this.internalJointImp.applyTranslation(x, y, z, asSeenBy);
-    }
-
-    @Override
-    public void applyRotationInRadians(Vector3 axis, double angleInRadians, ReferenceFrame asSeenBy) {
-      this.internalJointImp.applyRotationInRadians(axis, angleInRadians, asSeenBy);
-    }
-
-    @Override
-    public boolean isPivotVisible() {
-      return internalJointImp.isPivotVisible();
-    }
-
-    @Override
-    public void setPivotVisible(boolean isPivotVisible) {
-      internalJointImp.setPivotVisible(isPivotVisible);
-    }
-
-    private JointImp internalJointImp;
-    private JointImp jointParentWrapper;
-    private final List<JointImp> jointChildrenWrapper = new ArrayList<>();
+    void popJoint(JointImp joint);
   }
+
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Construction
+  // ════════════════════════════════════════════════════════════════════════════
 
   public JointedModelImp(A abstraction, JointImplementationAndVisualDataFactory<R> factory) {
     this.abstraction = abstraction;
-    this.factory = factory;
-    this.visualData = this.factory.createVisualData();
-    this.sgScalable = null;
+    this.resourceBinder = new JointedModelResourceBinder<>(factory);
+    this.visualManager = new JointedModelVisualManager<>(factory, null);
+    this.hierarchyManager = new JointHierarchyManager<>(this.resourceBinder);
 
-    Map<JointId, JointImp> jointMap = createJointImps();
-    //Make all the joint wrappers and put them in the map
-    for (Map.Entry<JointId, JointImp> entry : jointMap.entrySet()) {
-      JointImpWrapper wrapper = new JointImpWrapper(this, entry.getValue());
-      mapIdToJoint.put(entry.getKey(), wrapper);
-    }
-    fillInJointArrays();
-    //Go through all the wrappers and link up the parent/child relationship
-    for (Map.Entry<JointId, JointImpWrapper> entry : mapIdToJoint.entrySet()) {
-      entry.getValue().setJointParent(mapIdToJoint.get(entry.getKey().getParent()));
-    }
+    this.hierarchyManager.buildJointHierarchy(this);
 
-    this.visualData.setSGParent(getSgComposite());
+    this.visualManager.attachToParent(getSgComposite());
 
-    for (JointImpWrapper entry : mapIdToJoint.values()) {
-      final AbstractTransformable sgJoint = entry.getSgComposite();
+    for (JointImp joint : this.hierarchyManager.getJointWrappers()) {
+      final AbstractTransformable sgJoint = joint.getSgComposite();
       if (sgJoint.getParent() == null) {
         sgJoint.setParent(getSgComposite());
       }
     }
 
-    for (Visual sgVisual : this.visualData.getSgVisuals()) {
+    for (Visual sgVisual : this.visualManager.getSgVisuals()) {
       putInstance(sgVisual);
     }
-    for (SimpleAppearance sgAppearance : this.visualData.getSgAppearances()) {
+    for (SimpleAppearance sgAppearance : this.visualManager.getSgAppearances()) {
       putInstance(sgAppearance);
     }
   }
 
-  private Map<JointId, JointImp> createJointImps() {
-    List<JointId> allJointIds = getAllJointIds();
-    Map<JointId, JointImp> jointMap = new HashMap<>();
-    for (JointId jointId : allJointIds) {
-      jointMap.put(jointId, this.createJointImplementation(jointId));
-    }
-    for (JointId jointId : allJointIds) {
-      JointImp jointImp = jointMap.get(jointId);
-      JointImp parentImp = jointMap.get(jointId.getParent());
-      jointImp.setJointParent(parentImp);
-      if (jointImp.getSgVehicle() == null) {
-        jointImp.setVehicle(parentImp);
-      }
-    }
-    return jointMap;
-  }
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Resource change orchestration
+  // ════════════════════════════════════════════════════════════════════════════
 
-  private List<JointId> getAllJointIds() {
-    List<JointId> allJointIds = new ArrayList<>();
-    JointedModelResource resource = getResource();
-    for (Field jointField : ReflectionUtilities.getPublicStaticFinalFields(resource.getClass(), JointId.class)) {
-      try {
-        JointId jointId = (JointId) jointField.get(null);
-        allJointIds.add(jointId);
-      } catch (IllegalAccessException iae) {
-        Logger.throwable(iae, jointField);
-      }
-    }
-    if (resource instanceof DynamicResource dynamicResource) {
-      allJointIds.addAll(Arrays.asList(dynamicResource.getModelSpecificJoints()));
-    }
-    //Handle joint arrays
-    for (JointArrayId arrayId : this.getJointArrayIds()) {
-      JointId[] jointArrayIds = this.factory.getJointArrayIds(this, arrayId);
-      Collections.addAll(allJointIds, jointArrayIds);
-    }
-    return allJointIds;
-  }
-
-  private void fillInJointArrays() {
-    for (JointArrayId arrayId : this.getJointArrayIds()) {
-      mapArrayIdToJointIdArray.put(arrayId, findJointsMatching(arrayId.getElementNamePattern()));
-    }
-  }
-
-  private JointId[] findJointsMatching(String prefix) {
-    return mapIdToJoint.keySet().stream()
-                       .filter(jointId -> jointId.toString().startsWith(prefix))
-                       .sorted(Comparator.comparing(JointId::toString))
-                       .toArray(JointId[]::new);
-  }
-
-  public JointId[] getJointIdArray(JointArrayId jointArrayId) {
-    return this.mapArrayIdToJointIdArray.get(jointArrayId);
-  }
-
-  //TODO: Do we need this? Why do we need this.getJointArrayIds()?
-  public JointArrayId[] getJointArrayIds() {
-    if (this.jointArrayIds == null) {
-      List<JointArrayId> jointArrayIdsList = ReflectionUtilities.getPublicStaticFinalInstances(this.getResource().getClass(), JointArrayId.class);
-      this.jointArrayIds = jointArrayIdsList.toArray(new JointArrayId[0]);
-    }
-    return this.jointArrayIds;
-  }
-
+  @SuppressWarnings("unchecked")
   public void setNewResource(JointedModelResource resource) {
     if (resource == this.getResource()) {
       return;
     }
-    Composite originalParent = this.visualData.getSGParent();
-    VisualData<?> oldVisualData = this.visualData;
+    Composite originalParent = this.visualManager.getVisualData().getSGParent();
+    VisualData<?> oldVisualData = this.visualManager.getVisualData();
     Dimension3 oldScale = this.getScale();
     InstanceProperty[] oldScaleProperties = this.getScaleProperties();
-    this.factory = (JointImplementationAndVisualDataFactory<R>) resource.getImplementationAndVisualFactory();
+
+    JointImplementationAndVisualDataFactory<R> newFactory =
+        (JointImplementationAndVisualDataFactory<R>) resource.getImplementationAndVisualFactory();
+    this.resourceBinder.updateFactory(newFactory);
+    this.visualManager.updateFactory(newFactory);
+
     float originalOpacity = this.opacity.getValue();
     Paint originalPaint = this.paint.getValue();
-    this.visualData = this.factory.createVisualData();
 
-    if (!mapIdToJoint.isEmpty()) {
-      updateSkeleton();
+    this.visualManager.replaceVisualData();
+
+    if (!this.hierarchyManager.isEmpty()) {
+      this.hierarchyManager.updateSkeleton(this);
     }
 
-    this.visualData.setSGParent(originalParent);
-    oldVisualData.setSGParent(null);
+    this.visualManager.attachToParent(originalParent);
+    this.visualManager.detachOldVisualData(oldVisualData);
     this.opacity.setValue(originalOpacity);
     this.paint.setValue(originalPaint);
 
@@ -421,69 +178,9 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     this.setScale(oldScale);
   }
 
-  private void updateSkeleton() {
-    Map<JointId, JointImp> newJoints = createJointImps();
-    matchNewDataToExistingJoints(newJoints);
-    //Make joint wrappers for new entries and put them in the map
-    for (Map.Entry<JointId, JointImp> entry : newJoints.entrySet()) {
-      if (!mapIdToJoint.containsKey(entry.getKey())) {
-        mapIdToJoint.put(entry.getKey(), new JointImpWrapper(this, entry.getValue()));
-      }
-    }
-    mapArrayIdToJointIdArray.clear();
-    fillInJointArrays();
-  }
-
-  public JointedModelResource getVisualResource() {
-    return this.factory.getResource();
-  }
-
-  public Iterable<JointImp> getJoints() {
-    final List<JointImp> rv = Lists.newLinkedList();
-    this.treeWalk(new TreeWalkObserver() {
-      @Override
-      public void pushJoint(JointImp joint) {
-        rv.add(joint);
-      }
-
-      @Override
-      public void handleBone(JointImp parent, JointImp child) {
-      }
-
-      @Override
-      public void popJoint(JointImp joint) {
-      }
-    });
-    return rv;
-  }
-
-  private void matchNewDataToExistingJoints(Map<JointId, JointImp> newJoints) {
-    List<JointId> toRemove = Lists.newLinkedList();
-    for (JointId oldJointId : mapIdToJoint.keySet()) {
-      if (newJoints.containsKey(oldJointId)) {
-        mapIdToJoint.get(oldJointId).replaceWithJoint(newJoints.get(oldJointId));
-      } else {
-        toRemove.add(oldJointId);
-      }
-    }
-    // Order from outer-most toward root to always remove a leaf
-    toRemove.sort(JointId::descendantComparison);
-    for (JointId id : toRemove) {
-      JointImpWrapper impToRemove = this.mapIdToJoint.remove(id);
-      AbstractTransformable sgJoint = impToRemove.getSgComposite();
-      if (!impToRemove.getJointChildren().isEmpty()) {
-        Logger.severe("Removing a joint with child joints. There is a problem with this resource and may lead to errors.");
-      }
-      impToRemove.setJointParent(null);
-      sgJoint.setParent(null);
-    }
-  }
-
-  public void setAllJointPivotsVisible(boolean isPivotVisible) {
-    for (Map.Entry<JointId, JointImpWrapper> jointEntry : this.mapIdToJoint.entrySet()) {
-      jointEntry.getValue().setPivotVisible(isPivotVisible);
-    }
-  }
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Delegations — ResourceBinder
+  // ════════════════════════════════════════════════════════════════════════════
 
   @Override
   public A getAbstraction() {
@@ -491,21 +188,46 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
   }
 
   public R getResource() {
-    return this.factory.getResource();
+    return this.resourceBinder.getResource();
   }
 
+  public JointedModelResource getVisualResource() {
+    return this.resourceBinder.getResource();
+  }
+
+  public UnitQuaternion getOriginalJointOrientation(JointId jointId) {
+    return this.resourceBinder.getOriginalJointOrientation(jointId);
+  }
+
+  public AffineMatrix4x4 getOriginalJointTransformation(JointId jointId) {
+    return this.resourceBinder.getOriginalJointTransformation(jointId);
+  }
+
+  protected final JointImp createJointImplementation(JointId jointId) {
+    return this.resourceBinder.createJointImplementation(this, jointId);
+  }
+
+  //TODO: Do we need this? Why do we need this.getJointArrayIds()?
+  public JointArrayId[] getJointArrayIds() {
+    return this.resourceBinder.getJointArrayIds();
+  }
+
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Delegations — VisualManager
+  // ════════════════════════════════════════════════════════════════════════════
+
   public VisualData<R> getVisualData() {
-    return this.visualData;
+    return this.visualManager.getVisualData();
   }
 
   @Override
   public final Visual[] getSgVisuals() {
-    return this.visualData.getSgVisuals();
+    return this.visualManager.getSgVisuals();
   }
 
   @Override
   protected final SimpleAppearance[] getSgPaintAppearances() {
-    return this.visualData.getSgAppearances();
+    return this.visualManager.getSgAppearances();
   }
 
   @Override
@@ -513,20 +235,70 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     return this.getSgPaintAppearances();
   }
 
+  @Override
+  protected InstanceProperty[] getScaleProperties() {
+    return this.visualManager.getScaleProperties();
+  }
+
+  @Override
+  public Dimension3 getScale() {
+    return this.visualManager.getScale();
+  }
+
+  @Override
+  public void setScale(Dimension3 scale) {
+    if (this.visualManager.hasSgScalable()) {
+      this.visualManager.setScaleViaSgScalable(scale);
+    } else {
+      this.visualManager.setScaleOnVisuals(scale);
+      this.hierarchyManager.setScaleOnJoints(scale);
+    }
+  }
+
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Delegations — HierarchyManager
+  // ════════════════════════════════════════════════════════════════════════════
+
   public JointImp getJointImplementation(JointId jointId) {
-    return this.mapIdToJoint.get(jointId);
+    return this.hierarchyManager.getJointImplementation(jointId);
   }
 
   //String based lookup for DynamicJointIds
   public JointImp getJointImplementation(String jointName) {
-    //TODO: Think about maintaining a map of names to joints. It would be double accounting and might be a pain though
-    for (Map.Entry<JointId, JointImpWrapper> entry : this.mapIdToJoint.entrySet()) {
-      if (entry.getKey().toString().equals(jointName)) {
-        return entry.getValue();
-      }
-    }
-    return null;
+    return this.hierarchyManager.getJointImplementation(jointName);
   }
+
+  public JointId[] getJointIdArray(JointArrayId jointArrayId) {
+    return this.hierarchyManager.getJointIdArray(jointArrayId);
+  }
+
+  public Iterable<JointImp> getJoints() {
+    return this.hierarchyManager.getJoints();
+  }
+
+  public void setAllJointPivotsVisible(boolean isPivotVisible) {
+    this.hierarchyManager.setAllJointPivotsVisible(isPivotVisible);
+  }
+
+  public void treeWalk(TreeWalkObserver observer) {
+    this.hierarchyManager.treeWalk(observer);
+  }
+
+  public void straightenOutJoints() {
+    this.hierarchyManager.straightenOutJoints();
+  }
+
+  public List<JointImp> getInclusiveListOfJointsBetween(JointImp jointA, JointImp jointB, List<Bone.Direction> directions) {
+    return this.hierarchyManager.getInclusiveListOfJointsBetween(jointA, jointB, directions, this);
+  }
+
+  public List<JointImp> getInclusiveListOfJointsBetween(JointId idA, JointId idB, List<Bone.Direction> directions) {
+    return this.getInclusiveListOfJointsBetween(this.getJointImplementation(idA), this.getJointImplementation(idB), directions);
+  }
+
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Offset helpers (use this as reference frame)
+  // ════════════════════════════════════════════════════════════════════════════
 
   protected Vector4 getFrontOffsetForJoint(JointImp jointImp) {
     AxisAlignedBox bbox = jointImp.getAxisAlignedMinimumBoundingBox(this);
@@ -540,48 +312,9 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     return new Vector4(point.x(), point.y(), point.z(), 1);
   }
 
-  public UnitQuaternion getOriginalJointOrientation(JointId jointId) {
-    return this.factory.getOriginalJointOrientation(jointId);
-  }
-
-  public AffineMatrix4x4 getOriginalJointTransformation(JointId jointId) {
-    return this.factory.getOriginalJointTransformation(jointId);
-  }
-
-  @Override
-  protected InstanceProperty[] getScaleProperties() {
-    if (this.sgScalable != null) {
-      return new InstanceProperty[] {this.sgScalable.scale};
-    } else {
-      return new InstanceProperty[] {this.visualData.getSgVisuals()[0].scale};
-    }
-  }
-
-  @Override
-  public Dimension3 getScale() {
-    if (this.sgScalable != null) {
-      return this.sgScalable.scale.getValue();
-    } else {
-      Matrix3x3 scale = this.visualData.getSgVisuals()[0].scale.getValue();
-      return new Dimension3(scale.getRight().x(), scale.getUp().y(), scale.getBackward().z());
-    }
-  }
-
-  @Override
-  public void setScale(Dimension3 scale) {
-    if (this.sgScalable != null) {
-      this.sgScalable.scale.setValue(scale);
-    } else {
-      Matrix3x3 m = scale.asScaleMatrix();
-
-      for (Visual sgVisual : this.visualData.getSgVisuals()) {
-        sgVisual.scale.setValue(m);
-      }
-      for (JointImp jointImp : this.mapIdToJoint.values()) {
-        jointImp.setScale(scale);
-      }
-    }
-  }
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Bounding box / Size
+  // ════════════════════════════════════════════════════════════════════════════
 
   public AxisAlignedBox getAxisAlignedMinimumBoundingBox(ReferenceFrame asSeenBy, boolean ignoreJointOrientations) {
     AffineMatrix4x4 trans = this.getTransformation(asSeenBy);
@@ -643,9 +376,9 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     updateCumulativeBound(rv, trans, true);
   }
 
-  protected final JointImp createJointImplementation(JointId jointId) {
-    return this.factory.createJointImplementation(this, jointId);
-  }
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Visualization
+  // ════════════════════════════════════════════════════════════════════════════
 
   private JointedModelVisualization visualization;
 
@@ -669,242 +402,16 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     }
   }
 
-  private List<JointImp> getRootJointImps() {
-    List<JointImp> rootJoints = new ArrayList<>();
-    for (Map.Entry<JointId, JointImpWrapper> entry : mapIdToJoint.entrySet()) {
-      if (entry.getKey().getParent() == null) {
-        rootJoints.add(entry.getValue());
-      }
-    }
-    return rootJoints;
-  }
-
-  public interface TreeWalkObserver {
-    void pushJoint(JointImp joint);
-
-    void handleBone(JointImp parent, JointImp child);
-
-    void popJoint(JointImp joint);
-  }
-
-  private void treeWalk(JointImp parentImp, TreeWalkObserver observer) {
-    if (parentImp != null) {
-      observer.pushJoint(parentImp);
-      R resource = this.getResource();
-      for (JointImp childImp : parentImp.getJointChildren()) {
-        if (childImp != null) {
-          observer.handleBone(parentImp, childImp);
-        }
-      }
-      observer.popJoint(parentImp);
-      for (JointImp childImp : parentImp.getJointChildren()) {
-        treeWalk(childImp, observer);
-      }
-    }
-  }
-
-  public void treeWalk(TreeWalkObserver observer) {
-    for (JointImp root : this.getRootJointImps()) {
-      this.treeWalk(root, observer);
-    }
-  }
-
-  private enum AddOp {
-    PREPEND {
-      @Override
-      public List<JointImp> add(List<JointImp> rv, JointImp joint, List<Bone.Direction> directions, Bone.Direction direction) {
-        rv.addFirst(joint);
-        if (directions != null) {
-          directions.addFirst(direction);
-        }
-        return rv;
-      }
-    }, APPEND {
-      @Override
-      public List<JointImp> add(List<JointImp> rv, JointImp joint, List<Bone.Direction> directions, Bone.Direction direction) {
-        rv.add(joint);
-        if (directions != null) {
-          directions.add(direction);
-        }
-        return rv;
-      }
-    };
-
-    public abstract List<JointImp> add(List<JointImp> rv, JointImp joint, List<Bone.Direction> directions, Bone.Direction direction);
-  }
-
-  private List<JointImp> updateJointsBetween(List<JointImp> rv, List<Bone.Direction> directions, JointImp joint, EntityImp ancestorToReach, AddOp addOp) {
-    if (joint != ancestorToReach) {
-      JointId parentId = joint.getJointId().getParent();
-      if (parentId != null) {
-        JointImp parent = this.getJointImplementation(parentId);
-        this.updateJointsBetween(rv, directions, parent, ancestorToReach, addOp);
-      }
-    }
-    Bone.Direction direction;
-    if (addOp == AddOp.APPEND) {
-      direction = Bone.Direction.DOWNSTREAM;
-    } else {
-      direction = Bone.Direction.UPSTREAM;
-    }
-    addOp.add(rv, joint, directions, direction);
-    return rv;
-  }
-
-  public List<JointImp> getInclusiveListOfJointsBetween(JointImp jointA, JointImp jointB, List<Bone.Direction> directions) {
-    assert jointA != null : this;
-    assert jointB != null : this;
-    List<JointImp> rv = Lists.newLinkedList();
-    if (jointA == jointB) {
-      //?
-      rv.add(jointA);
-      directions.add(Bone.Direction.DOWNSTREAM);
-      //      throw new RuntimeException( "To Gazi: Please ensure that direction is correct in this case." );
-    } else {
-      if (jointA.isDescendantOf(jointB)) {
-        this.updateJointsBetween(rv, directions, jointA, jointB, AddOp.PREPEND);
-      } else if (jointB.isDescendantOf(jointA)) {
-        this.updateJointsBetween(rv, directions, jointB, jointA, AddOp.APPEND);
-      } else {
-        //It shouldn't even use the joint on which direction is changed (the common ancestor)
-        //that's what the below call does
-        this.updateJointsUpToAndExcludingCommonAncestor(rv, directions, jointA, jointB);
-      }
-    }
-    return rv;
-  }
-
-  private void updateJointsUpToAndExcludingCommonAncestor(List<JointImp> rvPath, List<Direction> rvDirections, JointImp jointA, JointImp jointB) {
-    List<JointImp> pathA = Lists.newLinkedList();
-    List<JointImp> pathB = Lists.newLinkedList();
-
-    List<Direction> directionsA = new ArrayList<Direction>();
-    List<Direction> directionsB = new ArrayList<Direction>();
-
-    this.updateJointsBetween(pathA, directionsA, jointA, this, AddOp.PREPEND);
-    this.updateJointsBetween(pathB, directionsB, jointB, this, AddOp.APPEND);
-
-    JointImp commonAncestor = null;
-
-    for (JointImp jointInA : pathA) {
-      if (pathB.contains(jointInA)) {
-        commonAncestor = jointInA;
-        break;
-      }
-    }
-
-    if (commonAncestor == null) {
-      throw new RuntimeException("Probably not connected with a chain.");
-    }
-
-    ListIterator<JointImp> pathAIterator = pathA.listIterator(pathA.size());
-    ListIterator<Direction> directionsAIterator = directionsA.listIterator(directionsA.size());
-    for (; pathAIterator.hasPrevious(); ) {
-      JointImp jointImp = pathAIterator.previous();
-      directionsAIterator.previous();
-
-      pathAIterator.remove();
-      directionsAIterator.remove();
-
-      if (jointImp == commonAncestor) {
-        break;
-      }
-    }
-
-    ListIterator<JointImp> pathBIterator = pathB.listIterator();
-    ListIterator<Direction> directionsBIterator = directionsB.listIterator();
-    for (; pathBIterator.hasNext(); ) {
-      JointImp jointImp = (JointImp) pathBIterator.next();
-      directionsBIterator.next();
-
-      pathBIterator.remove();
-      directionsBIterator.remove();
-
-      if (jointImp == commonAncestor) {
-        break;
-      }
-    }
-
-    rvPath.addAll(pathA);
-    rvPath.addAll(pathB);
-    rvDirections.addAll(directionsA);
-    rvDirections.addAll(directionsB);
-  }
-
-  public List<JointImp> getInclusiveListOfJointsBetween(JointId idA, JointId idB, List<Bone.Direction> directions) {
-    return this.getInclusiveListOfJointsBetween(this.getJointImplementation(idA), this.getJointImplementation(idB), directions);
-  }
-
-  private static class JointData {
-    private final JointImp jointImp;
-    private final UnitQuaternion q0;
-    private final UnitQuaternion q1;
-
-    public JointData(JointImp jointImp) {
-      this.jointImp = jointImp;
-      this.q0 = this.jointImp.getLocalOrientation().asUnitQuaternion();
-      UnitQuaternion q = this.jointImp.getOriginalOrientation();
-      this.q1 = ((q == null) || this.q0.isAlignedWith(q)) ? null : q;
-    }
-
-    //    public JointImp getJointImp() {
-    //      return this.jointImp;
-    //    }
-    //    public edu.cmu.cs.dennisc.math.UnitQuaternion getQ0() {
-    //      return this.q0;
-    //    }
-    //    public edu.cmu.cs.dennisc.math.UnitQuaternion getQ1() {
-    //      return this.q1;
-    //    }
-    public void setPortion(double portion) {
-      if (this.q1 != null) {
-        this.jointImp.setLocalOrientationOnly(this.q0.interpolate(this.q1, portion).asMatrix3x3());
-      } else {
-        //System.err.println( "skipping: " + this.jointImp );
-      }
-    }
-
-    public void epilogue() {
-      if (this.q1 != null) {
-        this.jointImp.setLocalOrientationOnly(this.q1.asMatrix3x3());
-      } else {
-        //System.err.println( "skipping: " + this.jointImp );
-      }
-    }
-  }
-
-  private static class StraightenTreeWalkObserver implements TreeWalkObserver {
-    private List<JointData> list = Lists.newLinkedList();
-
-    @Override
-    public void pushJoint(JointImp jointImp) {
-      list.add(new JointData(jointImp));
-    }
-
-    @Override
-    public void handleBone(JointImp parent, JointImp child) {
-    }
-
-    @Override
-    public void popJoint(JointImp joint) {
-    }
-  }
-
-  public void straightenOutJoints() {
-    StraightenTreeWalkObserver treeWalkObserver = new StraightenTreeWalkObserver();
-    this.treeWalk(treeWalkObserver);
-    for (JointData jointData : treeWalkObserver.list) {
-      jointData.epilogue();
-    }
-  }
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Animation
+  // ════════════════════════════════════════════════════════════════════════════
 
   public void animateStraightenOutJoints(double duration, Style style) {
     duration = adjustDurationIfNecessary(duration);
     if (EpsilonUtilities.isWithinReasonableEpsilon(duration, RIGHT_NOW)) {
       this.straightenOutJoints();
     } else {
-      final StraightenTreeWalkObserver treeWalkObserver = new StraightenTreeWalkObserver();
-      this.treeWalk(treeWalkObserver);
+      final List<JointHierarchyManager.JointData> jointDataList = hierarchyManager.collectStraightenData();
       class StraightenOutJointsAnimation extends DurationBasedAnimation {
         public StraightenOutJointsAnimation(double duration, Style style) {
           super(duration, style);
@@ -916,14 +423,14 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
 
         @Override
         protected void setPortion(double portion) {
-          for (JointData jointData : treeWalkObserver.list) {
+          for (JointHierarchyManager.JointData jointData : jointDataList) {
             jointData.setPortion(portion);
           }
         }
 
         @Override
         protected void epilogue() {
-          for (JointData jointData : treeWalkObserver.list) {
+          for (JointHierarchyManager.JointData jointData : jointDataList) {
             jointData.epilogue();
           }
         }
@@ -945,11 +452,12 @@ public abstract class JointedModelImp<A extends SJointedModel, R extends Jointed
     program.perform(new PoseAnimation(duration, style, this, pose), null);
   }
 
+  // ════════════════════════════════════════════════════════════════════════════
+  //  Fields
+  // ════════════════════════════════════════════════════════════════════════════
+
   private final A abstraction;
-  private final Scalable sgScalable;
-  private final Map<JointId, JointImpWrapper> mapIdToJoint = Maps.newHashMap();
-  private final Map<JointArrayId, JointId[]> mapArrayIdToJointIdArray = Maps.newHashMap();
-  private JointArrayId[] jointArrayIds = null;
-  private JointImplementationAndVisualDataFactory<R> factory;
-  private VisualData<R> visualData;
+  final JointedModelResourceBinder<R> resourceBinder;
+  final JointHierarchyManager<R> hierarchyManager;
+  final JointedModelVisualManager<R> visualManager;
 }
