+++ core/story-api/src/main/java/org/lgna/story/implementation/JointedModelResourceBinder.java
@@ -1,0 +1,134 @@
+/*******************************************************************************
+ * Copyright (c) 2006, 2015, Carnegie Mellon University. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * 3. Products derived from the software may not be called "Alice", nor may
+ *    "Alice" appear in their name, without prior written permission of
+ *    Carnegie Mellon University.
+ *
+ * 4. All advertising materials mentioning features or use of this software must
+ *    display the following acknowledgement: "This product includes software
+ *    developed by Carnegie Mellon University"
+ *
+ * 5. The gallery of art assets and animations provided with this software is
+ *    contributed by Electronic Arts Inc. and may be used for personal,
+ *    non-commercial, and academic use only. Redistributions of any program
+ *    source code that utilizes The Sims 2 Assets must also retain the copyright
+ *    notice, list of conditions and the disclaimer contained in
+ *    The Alice 3.0 Art Gallery License.
+ *
+ * DISCLAIMER:
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND.
+ * ANY AND ALL EXPRESS, STATUTORY OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY,  FITNESS FOR A
+ * PARTICULAR PURPOSE, TITLE, AND NON-INFRINGEMENT ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE AUTHORS, COPYRIGHT OWNERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING FROM OR OTHERWISE RELATING TO
+ * THE USE OF OR OTHER DEALINGS WITH THE SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *******************************************************************************/
+
+package org.lgna.story.implementation;
+
+import edu.cmu.cs.dennisc.java.lang.reflect.ReflectionUtilities;
+import edu.cmu.cs.dennisc.java.util.logging.Logger;
+import org.alice.math.immutable.AffineMatrix4x4;
+import org.alice.math.immutable.UnitQuaternion;
+import org.lgna.story.implementation.JointedModelImp.JointImplementationAndVisualDataFactory;
+import org.lgna.story.implementation.JointedModelImp.VisualData;
+import org.lgna.story.resources.DynamicResource;
+import org.lgna.story.resources.JointArrayId;
+import org.lgna.story.resources.JointId;
+import org.lgna.story.resources.JointedModelResource;
+
+import java.lang.reflect.Field;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+
+/**
+ * Encapsulates factory delegation, resource queries, isSims() check,
+ * and reflection-based joint ID discovery extracted from JointedModelImp.
+ */
+class JointedModelResourceBinder<R extends JointedModelResource> {
+
+  private JointImplementationAndVisualDataFactory<R> factory;
+  private JointArrayId[] jointArrayIds;
+
+  JointedModelResourceBinder(JointImplementationAndVisualDataFactory<R> factory) {
+    this.factory = Objects.requireNonNull(factory, "factory");
+  }
+
+  R getResource() {
+    return factory.getResource();
+  }
+
+  boolean isSims() {
+    return factory.isSims();
+  }
+
+  List<JointId> getAllJointIds() {
+    List<JointId> allJointIds = new ArrayList<>();
+    JointedModelResource resource = getResource();
+    for (Field jointField : ReflectionUtilities.getPublicStaticFinalFields(resource.getClass(), JointId.class)) {
+      try {
+        JointId jointId = (JointId) jointField.get(null);
+        allJointIds.add(jointId);
+      } catch (IllegalAccessException iae) {
+        Logger.throwable(iae, jointField);
+      }
+    }
+    if (resource instanceof DynamicResource dynamicResource) {
+      allJointIds.addAll(Arrays.asList(dynamicResource.getModelSpecificJoints()));
+    }
+    return allJointIds;
+  }
+
+  JointArrayId[] getJointArrayIds() {
+    if (this.jointArrayIds == null) {
+      List<JointArrayId> list = ReflectionUtilities.getPublicStaticFinalInstances(
+          getResource().getClass(), JointArrayId.class);
+      this.jointArrayIds = list.toArray(new JointArrayId[0]);
+    }
+    return this.jointArrayIds;
+  }
+
+  UnitQuaternion getOriginalJointOrientation(JointId jointId) {
+    return factory.getOriginalJointOrientation(jointId);
+  }
+
+  AffineMatrix4x4 getOriginalJointTransformation(JointId jointId) {
+    return factory.getOriginalJointTransformation(jointId);
+  }
+
+  void updateFactory(JointImplementationAndVisualDataFactory<R> newFactory) {
+    this.factory = Objects.requireNonNull(newFactory, "newFactory");
+    this.jointArrayIds = null;
+  }
+
+  JointImp createJointImplementation(JointedModelImp<?, R> owner, JointId jointId) {
+    return factory.createJointImplementation(owner, jointId);
+  }
+
+  JointId[] getJointArrayIdsFromFactory(JointedModelImp<?, R> owner, JointArrayId arrayId) {
+    return factory.getJointArrayIds(owner, arrayId);
+  }
+
+  VisualData<R> createVisualData() {
+    return factory.createVisualData();
+  }
+}
